requirements.txt
bandit==1.6.2
beautifulsoup4==4.9.3
Flask==2.0.1
grpcio==1.38.0
html5lib==1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mechanize==0.4.5
flask_script
flask_migrate
flask_session
flask_login
requests
psycopg2-binary
lxml
pandas
boto3
pillow
google-cloud-datastore
gunicorn